feat: query package
