+++ packages/query/ambient.d.ts
@@ -1,0 +1,10 @@
+declare function $<T extends Element = HTMLElement>(selector: string): T;
+
+declare namespace $ {
+	function all<T extends Element = HTMLElement>(selector: string): Array<T>;
+	function optional<T extends Element = HTMLElement>(
+		selector: string,
+	): T | undefined;
+}
+
+declare function ready(callback: () => void): undefined | (() => void);
